requirements.txt
streamlit>=1.28.0
pathlib>=1.0.1
streamlit>=1.28.0
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
tabulate>=0.9.0
crewai-tools>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
streamlit==1.29.0
anthropic==0.18.1
openai==1.12.0
google-generativeai==0.3.2
python-dotenv==1.0.0
sqlalchemy==2.0.25
pydantic==2.5.3
PyPDF2==3.0.1
tiktoken 
rapidfuzz 
sentence-transformers
openai
anthropic
python-dotenv
streamlit 
crewai 
langchain-anthropic 
anthropic 
pandas
numpy
requests
pdfplumber
PyPDF2
